requirements.txt
appnope==0.1.2
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
dataclasses==0.8
decorator==5.0.9
Flask==2.0.1
future==0.18.2
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.6.4
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.1
line-bot-sdk==1.20.0
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
zipp==3.5.0
